finish step counter sensor and temperature sensor!

+++ TerveysHelppi/app/src/main/java/com/example/terveyshelppi/Service/RoomDB/UserData.kt
@@ -15,7 +15,7 @@ data class UserData(
     val heartRate: Int,
     val totalDistance: Int,
     val totalCalories: Int,
-    val totalSteps: Int,
+    val totalSteps: Double,
     val totalHours: Int
 
 ) {
